Fixed everything :)
The Spring Boot application should now run without a hitch.
Also implemented a temporary RDS endpoint, just to get us to stop using H2.
On top of this, I got rid of the h2 dependency
