+++ revature-cart-backend/src/main/java/com/revature/cart/service/CartService.java
@@ -1,11 +1,10 @@
-package com.revature.service;
+package com.revature.cart.service;
+
 import java.util.List;
-import java.util.Optional;
 
 import com.revature.cart.model.Cart;
-import com.revature.cart.model.CartItem;
-public interface CartItemService {
-	
+
+public interface CartService {
 	
 	// CREATE
 	
@@ -14,51 +13,37 @@ public interface CartItemService {
 	 * @param item The object of the model passed in to save
 	 * @return if saved, returns the saved item, if not returns null
 	 */
-	public CartItem createCartItem(CartItem item);
+	public Cart createCart(Cart cart);
 	
 	// READ
-
+	
 	/**
 	 * Gets all the cart items
-	 * @return List<CartItem> the list of items retrieved by JPA
+	 * @return List<Cart> the list of items retrieved by JPA
 	 */
-	public List<CartItem> getAllCartItems();
+	public List<Cart> getAllCarts();
 	
 	/**
 	 * Gets a cart item based on a given ID
 	 * @param id the id of the cart item
 	 * @return the cart item returned by JPA
 	 */
-	public CartItem getCartItemById(int id);
-	
+	public Cart getCartById(int id);
 	
 	// UPDATE
-	
 	/**
 	 * Updates a specific cart item by id
 	 * @param id the id of the item to be updated
 	 * @return 
 	 */
-	public CartItem updateCartItemById(int id);
+	public Cart updateCartById(int id);
+	public Cart updateCart(Cart cart);
 	
-	public CartItem updateCartItem(CartItem item);
-	
-	// DELETE 
-	
+	// DELETE
 	/**
 	 * Deletes the cart item
 	 * @param id the id of the item to be deleted
 	 */
-	public void deleteCartItemById(int id);
-	
-	public void deleteCartItem(CartItem item);
-
-	
-
-	
-
-	
-
-
-
+	public void deleteCartById(int id);
+	void deleteCart(Cart cart);
 }
