+++ revature-cart-backend/src/main/resources/application.properties
@@ -1,8 +1,12 @@
-
-spring.application.name=rss
-spring.jpa.show-sql=true
-spring.jpa.properties.hibernate.format_sql=true
+spring.application.name=revature-cart-backend
 spring.jpa.hibernate.ddl-auto=update
 spring.h2.console.enabled=true
 spring.h2.console.path=/h2
 server.port=9999
+
+spring.datasource.url=jdbc:oracle:thin:@myfirstorcl.caacxjx4qkkr.us-east-2.rds.amazonaws.com:1521:orcl
+spring.datasource.username=test2
+spring.datasource.password=test2
+
+spring.jpa.show-sql=true
+spring.jpa.properties.hibernate.format_sql=true
